Fixes and refactoring for TexMobject
